# perf: cache empty buffer creation