@wozeparrot 2023-08-26
hmm, `create_lazybuffer` does some caching already, can that caching just be used? Tho in `create_lazybuffer` there is a thing to prevent CONST from getting cached, but I think it might be unnecessary? cuz CONSTs don't change unlike FROM and RAND