@wozeparrot 2023-08-26
i'm not sure if empty can be cached cuz they imply a memory allocation and there is disktensors which use empty to read.

what was the issue with caching consts?